refactor: Convertor 클래스를 Util 클래스로 통합
